moved test_quandl() to test_quandl.py in test dir
